backend/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
python-multipart

# Authentication
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv

# Statistical & Data Processing
numpy
scipy
pandas
scikit-learn
statsmodels

# Data Export
openpyxl

# Database
sqlalchemy
xlsxwriter

# Validation & Testing
pydantic
email-validator

# Utilities
python-dateutil
pytz
